backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
twilio
requests
